Text_to_SQL_LLM_App_along_with_Quering_SQL_database_using_GeminiPro/requirements.txt
google-generativeai
python-dotenv
streamlit
langchain
PyPDF2
chromadb
faiss-cpu
pdf2image
tiktoken
protobuf